requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
mysql-connector-python==9.0.0
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
